+++ src/imio/smartweb/policy/__init__.py
@@ -1,2 +1,9 @@
 # -*- coding: utf-8 -*-
-"""Init and utils."""
+
+from AccessControl import allow_module, allow_class, allow_type
+from requests.models import Response
+
+# Useful for code used in RestrictedPython (collective.themefragments)
+allow_module("requests")
+allow_class(Response)
+allow_type(type("requests.models.Response"))